pyproject.toml
[project]
name = "infra-guard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "DevHusnainAi", email = "syedhussnaintirmizi@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.134.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[project.scripts]
infra-guard = "infra_guard:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.3",
]
